+++ tests/similarity/test_cosine_greedy.py
@@ -67,9 +67,9 @@ def test_cosine_greedy_matrix(symmetric):
     spectrum_2 = builder.with_mz(np.array([110, 190, 290], dtype="float")).with_intensities(
         np.array([0.5, 0.2, 1.0], dtype="float")).build()
 
-    spectrums = [spectrum_1, spectrum_2]
+    spectra = [spectrum_1, spectrum_2]
     cosine_greedy = CosineGreedy()
-    scores = cosine_greedy.matrix(spectrums, spectrums, is_symmetric=symmetric)
+    scores = cosine_greedy.matrix(spectra, spectra, is_symmetric=symmetric)
 
     assert scores[0][0][0] == pytest.approx(scores[1][1][0], 0.000001), "Expected different cosine score."
     assert scores[0][0]["score"] == pytest.approx(scores[1][1]["score"], 0.000001), \
@@ -85,6 +85,6 @@ def test_cosine_greedy_matrix_unsymmetric_error():
 
     spectrum_2 = builder.with_mz(np.array([110, 190, 290], dtype="float")).with_intensities(
         np.array([0.5, 0.2, 1.0], dtype="float")).build()
-    
+
     with pytest.raises(ValueError, match="unequal number of spectra"):
         CosineGreedy().matrix([spectrum_1, spectrum_2], [spectrum_2], is_symmetric=True)