+++ tests/similarity/test_metadata_match.py
@@ -6,7 +6,7 @@ from tests.builder_Spectrum import SpectrumBuilder
 
 
 @pytest.fixture
-def spectrums():
+def spectra():
     metadata1 = {"instrument_type": "orbitrap",
                  "retention_time": 100.}
     metadata2 = {"instrument_type": "qtof",
@@ -22,31 +22,31 @@ def spectrums():
     return [s1, s2, s3, s4]
 
 
-def test_metadata_match_strings(spectrums):
+def test_metadata_match_strings(spectra):
     """Test basic metadata matching between string entries."""
-    references = spectrums[:2]
-    queries = spectrums[2:]
+    references = spectra[:2]
+    queries = spectra[2:]
 
     similarity_score = MetadataMatch(field="instrument_type")
     scores = calculate_scores(references, queries, similarity_score)
     assert np.all(scores.scores.to_array() == [[1, 0], [0, 0]]), "Expected different scores."
 
 
-def test_metadata_match_strings_pair(spectrums):
+def test_metadata_match_strings_pair(spectra):
     """Test basic metadata matching between string entries."""
     similarity_score = MetadataMatch(field="instrument_type")
-    score = similarity_score.pair(spectrums[0], spectrums[1])
+    score = similarity_score.pair(spectra[0], spectra[1])
     assert score == np.array(False, dtype=bool), "Expected different score."
-    score = similarity_score.pair(spectrums[0], spectrums[3])
+    score = similarity_score.pair(spectra[0], spectra[3])
     assert score == np.array(False, dtype=bool), "Expected different score."
-    score = similarity_score.pair(spectrums[0], spectrums[2])
+    score = similarity_score.pair(spectra[0], spectra[2])
     assert score == np.array(True, dtype=bool), "Expected different score."
 
 
-def test_metadata_match_strings_wrong_method(spectrums, caplog):
+def test_metadata_match_strings_wrong_method(spectra, caplog):
     """Test basic metadata matching between string entries."""
-    references = spectrums[:2]
-    queries = spectrums[2:]
+    references = spectra[:2]
+    queries = spectra[2:]
 
     similarity_score = MetadataMatch(field="instrument_type", matching_type="difference")
     scores = calculate_scores(references, queries, similarity_score)
@@ -55,12 +55,12 @@ def test_metadata_match_strings_wrong_method(spectrums, caplog):
     assert msg in caplog.text
 
 
-def test_metadata_match_numerical_pair(spectrums):
+def test_metadata_match_numerical_pair(spectra):
     """Test basic metadata matching between string entries."""
     similarity_score = MetadataMatch(field="retention_time",
                                      matching_type="difference",
                                      tolerance=0.6)
-    score = similarity_score.pair(spectrums[0], spectrums[1])
+    score = similarity_score.pair(spectra[0], spectra[1])
     assert score == 1, "Expected different score."
 
 
@@ -70,10 +70,10 @@ def test_metadata_match_numerical_pair(spectrums):
     [10.0, [[1, 1], [1, 1]]],
     [0.1, [[0, 0], [0, 0]]]
 ])
-def test_metadata_match_numerical(spectrums, tolerance, expected):
+def test_metadata_match_numerical(spectra, tolerance, expected):
     """Test basic metadata matching between numerical entries."""
-    references = spectrums[:2]
-    queries = spectrums[2:]
+    references = spectra[:2]
+    queries = spectra[2:]
 
     similarity_score = MetadataMatch(field="retention_time",
                                      matching_type="difference", tolerance=tolerance)
@@ -81,10 +81,10 @@ def test_metadata_match_numerical(spectrums, tolerance, expected):
     assert np.all(scores.scores.to_array().tolist() == expected), "Expected different scores."
 
 
-def test_metadata_match_invalid_array_type(spectrums):
+def test_metadata_match_invalid_array_type(spectra):
     """Test value error if array_type is not 'numpy' or 'sparse' in metadata matching."""
-    references = spectrums[:2]
-    queries = spectrums[2:]
+    references = spectra[:2]
+    queries = spectra[2:]
 
     similarity_score = MetadataMatch(field="instrument_type")